Added literal_map structure
Exploiting that we represent literals as signed integers,
I made a simple "map" structure that maps signed integers
of a fixed range to any element. I.e., an array.

This may be something of a premature optimization, but it's
conceptually simple, and I'm going to want to instrument
basically every method anyways, even if I used a built-in std::map.