Add cnf_table structure
Added a structure that is hopefully a (reasonable) representation
of a CNF. The actual clauses are encoded in a single contiguous
chunk, and the clauses are represented by a pair of [start, finish)
offsets.

I'm deliberately trying to make it so that a clause, and a clause*
are both treated equivalently. Again, maybe a slight concern
for premature optimization, but I think being able to have a
clause_iterator act as a clause "handle" would be very valuable.