Added a simple DIMACS cnf parser module
I've been holding onto this scrap code in older versions for
a very long time. This is finally a simple object I should be
able to link to and easily parse in DIMACS cnf formatted files.

Note that it is not entirely correct. I should revisit this one
day to make sure it's robust to hand valid-but-weird-looking
CNF files.