Added basic watched_literals structure
To implement the Chaff watched-literals scheme we need some helper
structures mapping literals to the clauses they watch, and clauses to
the pair of literals watching them. This change provides a starting
point.

Design decision: the main propagation "reasoning" takes an assignment as
a parameter. In this way we're saying that if we backtrack, the watched
literal structure is still valid, independent of the state of the
assignment. This is the purported advantage of the watched literal
scheme, or at least one advantage.

The main prop. method is completely untested. This is just a starting
point where all the data structures are at least initialized, and I can
start getting some necessary trace statements up and running. The design
of the prop method is inspired from reading a paper, "cache
efficient..." something something, proper citation TBD.