Verified backtracking invariant, perf improvement
I measured and found that (on a sample input), about a third of the
total time was spent simply copying the watch_literals structure over
the course of backtracking. This was deliberate, I anticipated problems
maintaining the back-track-ability invariant of the watched literals, so
I took that hit to get things up and running.

By simply changing the parameters from by-val to by-ref I was able to
see if the backtracking invariant worked, and it seems to hold! At
least, over all the AIM input sets of size 50 variables. Plus, things
get much much faster. Now, 2/3rds of the runtime is dominated by adding
and removing to the red-black trees. I know those sets are horribly
expensive, but I'm glad for the confirmation.

(All measurements are dead-simple at this point: simple valgrind with
callgrind, and then kcachegrind to see the call tree.)