Improved basic data structures
I ripped out the std::map, std::set usage, and consequently saw a
substantive performance improvement.

It's still not done, and there are rough edges in the implementations --
there are iterators, const, method details that aren't fully correct,
but it's not worth the time in light of larger implementation details.