Initial commit for clause learning framework
The first rough outline of the code needed to get a clause learning
framework. The main dpll procedure is no longer recursive, and instead
there's a quite primitive "assignment" structure that's used to keep
track of what literal assignments we should use next.

I've started also recording the information necessary for the
implication graph, which is likely the cause for the big performance
regression in this code. I'm not yet concerned about performance.

The next logical steps are to get an actual clause-learning routine
plugged in. That will also necessitate our ability to grow our core data
structures.