NCB (Non-chronological backtracking) implemented
As detailed in section 4.3 of the D&N paper. On my machine, this reduces
the aim-*50* benchmark from 2.4-ish seconds to 1.4-ish seconds. That's a
huge improvement.

There are some small complications: if the newparent clause is a unit
clause it's not obvious if things are benificial. Also, my
ChooseNextVariable function is incredibly simple, which would lead the
NCB into infinite loops in some instances. I work-around this by
preventing NCB in cases where we seem to be re-approaching our original
decision sequence.

Note that even with these hacky gaps, we get a real performance
improvement! Very nice.