Implemented extremely basic BCP
Using the most straightforward "linear" search for unit clauses
imaginable, I implemented the outline of BCP as recommended in the D&N
paper.

Of course I should really implement watched literals over this, but even
so I get a ridiculous speedup: 3x or so on the aim-*50* benchmark, and
now the aim-*100* benchmark is in tractable reach, about 3.8 seconds.

General log: It's really neat to see all this come together, but there
are still some open questions. I'm frustrated that the literature does
not seem to discuss soundness/completeness in the context of NCB, but
I've come to terms with my hackneyed work-around for now. Either way I
of course intend to implement a competitive heuristic, but it's not
obvious (to me) that any such heuristic is sound or complete.