Probably patched NCB.
I had a very subtle off-by-one sort-of error. The main intuition I was
previously missing: the real goal is to move the implication as "high
up" the decision sequence as possible. So, if we're at level 8 and find
we have a new implication, (i.e., a "right" move), but it really could
have been implied at level 4, we change things to force that implication
at level 4.

It was hard to realize that was what I was missing until I took the time
to brutally cut down some repros. So those are the tests I've checked in
with this, though they're not unit tests in the usual sense.