Refactor phase 1
I wrote in a whole bunch of comments and assertions to solidify my
understanding of this really quite subtle algorithm. I did some ad-hoc
performance testing to make sure I wasn't accidentally changing any
algorithms.

The trace statement format is settling into a more consistent one.
Hopefully it may be feasible to really analyze a full-fledged execution,
both for correctness and performance purposes. Right now things seem
good.

Other refactoring notes:
1) decision_sequence has double-size arrays than needed. Not a huge
problem, but obviously shouldn't be there.
2) NCB especially needs cleanup, and a bit of CDB.
3) Open question: right now a lot of the discrete phases are all jammed
into the main function. I like having everything there and explicit
(remember, this is for expositional purposes, not software engineering),
but it's not obvious if that makes it better or worse for studying. It
is certainly making it easy for me to "audit" the code, at least...