Refactor 2: cleaning up NCB
The NCB code continues to befuddle me. I have revisited it and cleaned
it up. I think the papers I'm referring to simply phrase it in a way
that is either wrong (unlikely but possible) or I'm constantly
misreading it (it's happened).

With this checkin, I think I have a comment explaining my understanding,
and a fairly neat set of code to implement NCB. CDB is the next and
final thing that I don't have a 100 percent understanding for.