Watched Literals
This brings in the core watched literals implementation. It successfully
improves the SSA benchmarks I have from 5 seconds to 0.5 seconds.

The implementation is fairly by-the-book. The conceptually trickiest
part is maintaining the validity of the implied literals cache in the
presence of NCB. Maybe I'm missing something obvious, but the approach I
settled on was to propogate units as much as possible, and in the case
of there being a conflict, quickly backtrack until just before hitting
that first conflict so that the NCB and other conflict-handling
reasoning can take over.

There is some leftover technical debt, mainly dead methods in helper
classes, but I've written a lot of assertions maintaining otherwise
implicit invariants that were revealed by implementing WL. More to come.