Refactor: CNF resizable, fewer small_set clauses
1) Excised some of the small_set<literal> "clauses" from the main loop,
and replaced with direct clause_iterators. This includes new_parent,
which means it is necessary for termination that we learn clauses.

2) The decision sequence no longer implicitly maintains a whole
permutation of possible literals, but instead accurately acts as a stack
of literals we're considering in our assignment.

3) The decision sequence Parent vector now stores clause_iterators
directly, setting the stage for a clause-based-heuristic.

4) The CNF table is now resizable, and can grow. This necessitated
implementing "on-resize" callbacks to update all data structures that
contain clause_iterators. We can now learn an indefinite number of
clauses. This callback system is likely setting the stage for a
clause-cleaning procedure, where we can remap clauses.

5) Some redundant helpers have been cleaned up, especially some
fundamental definitions, and excessive templatization has been scaled
back.