Glue Clause, first stab at implementation
This is the first implementation at A&S's glue clause heuristic for
clause-database cleaning. The biggest machinery is that the CNF table
can now "remap" clauses, changing the order of clauses in the DB. By
rearranging, and then decrementing the number of clauses, we essentially
allow for clause removal.

There's a few rough edges, but against the uuf100 inputs, everything is
about twice as fast (maybe not quite that much). 5 minutes instead of 9
minutes and so on.

Interestingly, the one factoring input I have, there doesn't seem to be
much clause learning.

I'm sure a lot of the algorithm can be cleaned up, I have a feeling some
of my pointer arithmetic can be made much more elegant. Additional work
to smooth this out would be to linearize the clause bodies in raw_data.
I assume that would help, but no measurements yet.

Another question: I've confirmed visually that I really am cleaning up
clauses with this algorithm, but is it behaving as I really intend, or
is there some subtle performance bug? This is a general concern, as all
these moving pieces get more and more complicated.