Small function tweak leads to 30% perf
I looked at perf ./sat < input/uuf100-, and found that basically 100% of
our time was spent determining if clauses were unsat. This is not what
is supposed to happen.

I haven't dug in to really examine the code-gen, but it's obvious that I
made a mistake with std::all_of. I played around with a few
alternatives, and the one here is the only with a perf improvement, and
it's a considerable improvement.