Full rewrite, now at version 0.3.0
I rewrote the main SAT solver. The helper data structures are
more-or-less the same. The design principle is that I really wanted to
be able to defend (explain) every facet of the implementation.

Some of the D&N decisions -- spooling out the BCP propogation, flipping
falsified literals rather than just backtracking, and more -- are ones I
couldn't fully explain the reasoning for. At this point I have the
experience to be able to take a step back and implement this newer,
working copy.

Note: a very subtle (i.e., buggy) data structure is the way I've
implemented watch literals. I don't have a fully formal reason, but
experimentally I determined that I needed (for unit clauses) to have the
other watching literal be the highest-decision-level literal.

There's a bit of cleanup to do, but now I have a base I'm comfortable
building on.