VSIDS added
That was the fastest implementation ever. Definitely improves
performance.

At this point, the main missing pieces are the restart strategy, the
trace statements, and the cleanup.

While there is a lot of "in-processing" missing, I think it would be
appropriate to use the (polished) implementation for an exposition on
the implementation of a "modern" SAT solver that undergrads may learn
about. I'm not trying to compete with CryptoMiniSat or anything.